tuvi-python/requirements.txt
lunarcalendar
fastapi
uvicorn
pydantic
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
python-multipart
python-dotenv
mysql-connector-python
